+++ packages/agtree/src/parser/misc/logical-expression.ts
@@ -135,6 +135,7 @@ const KNOWN_VARIABLES_MAP = new Map<string, number>([
     ['adguard_ext_edge', 21],
     ['adguard_ext_opera', 22],
     ['adguard_ext_android_cb', 23],
+    // TODO: Add 'adguard_ext_chromium_mv3' to the list
 ]);
 
 /**