+++ packages/tswebextension/src/lib/mv2/background/request/request-context-storage.ts
@@ -7,6 +7,9 @@ import type { ContentType } from '../../../common/request-type';
 import type { ParsedCookie } from '../../../common/cookie-filtering/parsed-cookie';
 import type { TabFrameRequestContext } from '../tabs/tabs-api';
 
+/**
+ * Request context state. It represents the current state of the request processing.
+ */
 export const enum RequestContextState {
     BeforeRequest = 'beforeRequest',
     BeforeSendHeaders = 'beforeSendHeaders',
@@ -32,7 +35,11 @@ export type RequestContext = TabFrameRequestContext & {
     eventId: string;
 
     state: RequestContextState;
-    timestamp: number; // webRequest event timestamp
+
+    /**
+     * WebRequest event timestamp, in milliseconds since the epoch.
+     */
+    timestamp: number;
     referrerUrl: string;
     contentType: ContentType;
     requestFrameId: number;